+++ app/src/main/java/net/kwmt27/githubsearch/presenter/search/ISearchPresenter.java
@@ -11,5 +11,4 @@ public interface ISearchPresenter extends IBaseActivityPresenter {
 
     String REPO_ENTITY_KEY = "repo_entity_key";
 
-    void onClickReloadButton();
 }
