+++ app/src/main/java/net/kwmt27/githubsearch/view/search/FragmentProgressCallback.java
@@ -5,7 +5,6 @@ public interface FragmentProgressCallback {
 
     void hideProgress();
 
-    void showError();
 
     void showNotFoundPageIfNeeded(boolean show);
 }
