+++ app/src/main/res/layout/activity_search.xml
@@ -73,13 +73,6 @@
 
 
     <include
-        layout="@layout/include_error_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/toolbar"
-        />
-
-    <include
         layout="@layout/include_progress_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
