+++ app/src/main/res/layout/fragment_search_result_list.xml
@@ -12,4 +12,11 @@
         android:scrollbars="vertical"
         />
 
+    <include
+        layout="@layout/include_error_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        />
+
+
 </RelativeLayout>
